:truck: Move express middleware to express folder
