+++ src/drivers/express/middleware/process-token.ts
